Refactor uninstall to add -purge flag (#47)
    * Refactor uninstall to only remove program data files when the `-purge` flag is passsed

+++ containers-toolkit/Public/BuildkitTools.psm1
@@ -59,13 +59,13 @@ function Install-Buildkit {
         # Check if Buildkit is alread installed
         $isInstalled = -not (Test-EmptyDirectory -Path $InstallPath)
 
-        $WhatIfMessage = "Buildkit will be installed at $InstallPath"
+        $WhatIfMessage = "Buildkit will be installed at '$InstallPath'"
         if ($isInstalled) {
-            $WhatIfMessage = "Buildkit will be uninstalled from and reinstalled at $InstallPath"
+            $WhatIfMessage = "Buildkit will be uninstalled from and reinstalled at '$InstallPath'"
         }
         if ($Setup) {
             <# Action when this condition is true #>
-            $WhatIfMessage = "Buildkit will be installed at $InstallPath and buildkitd service will be registered and started"
+            $WhatIfMessage = "Buildkit will be installed at '$InstallPath' and buildkitd service will be registered and started"
         }
     }
 
@@ -73,7 +73,7 @@ function Install-Buildkit {
         if ($PSCmdlet.ShouldProcess($env:COMPUTERNAME, $WhatIfMessage)) {
             # Check if tool already exists at specified location
             if ($isInstalled) {
-                $errMsg = "Buildkit already exists at $InstallPath or the directory is not empty"
+                $errMsg = "Buildkit already exists at '$InstallPath' or the directory is not empty."
                 Write-Warning $errMsg
 
                 # Uninstall if tool exists at specified location. Requires user consent
@@ -95,13 +95,13 @@ function Install-Buildkit {
 
             # Download files
             $downloadParams = @{
-                ToolName = "Buildkit"
-                Repository = "moby/buildkit"
-                Version = $Version
-                OSArchitecture = $OSArchitecture
-                DownloadPath = $DownloadPath
+                ToolName           = "Buildkit"
+                Repository         = "moby/buildkit"
+                Version            = $Version
+                OSArchitecture     = $OSArchitecture
+                DownloadPath       = $DownloadPath
                 ChecksumSchemaFile = "$ModuleParentPath\Private\schemas\in-toto.sbom.schema.json"
-                FileFilterRegEx = $null
+                FileFilterRegEx    = $null
             }
             $downloadParamsProperties = [FileDownloadParameters]::new(
                 $downloadParams.ToolName,
@@ -145,7 +145,7 @@ function Install-Buildkit {
             }
 
             # Show buildkit binaries help
-            Get-ChildItem -Path "C:\Program Files\buildkit\bin" | ForEach-Object {
+            (Get-command -Name "buildctl","buildkitd" -ErrorAction SilentlyContinue) | ForEach-Object {
                 $executable = $_.Name
                 # Remove extension from executable
                 $commandName = $executable -replace ".exe", ""
@@ -339,7 +339,10 @@ function Uninstall-Buildkit {
     )]
     param(
         [parameter(HelpMessage = "BuildKit path")]
-        [String]$Path,
+        [String]$Path = "$ENV:ProgramFiles\Buildkit",
+
+        [parameter(HelpMessage = "Remove all program data for Buildkit")]
+        [Switch] $Purge,
 
         [parameter(HelpMessage = "Bypass confirmation to uninstall BuildKit")]
         [Switch] $Force
@@ -351,28 +354,34 @@ function Uninstall-Buildkit {
             $Path = Get-DefaultInstallPath -Tool $tool
         }
 
-        $WhatIfMessage = "Buildkit will be uninstalled from $path and buildkitd service will be stopped and unregistered"
+        $WhatIfMessage = "Buildkit will be uninstalled from '$path' and buildkitd service will be stopped and unregistered."
+        if ($Purge) {
+            $WhatIfMessage += " Buildkit program data will also be removed."
+        }
+        else {
+            $WhatIfMessage += " Buildkit program data won't be removed. To remove program data, run 'Uninstall-Buildkit' command without -Purge flag."
+        }
     }
 
     process {
         if ($PSCmdlet.ShouldProcess($env:COMPUTERNAME, $WhatIfMessage)) {
-            if (Test-EmptyDirectory -Path $path) {
-                Write-Output "$tool does not exist at $Path or the directory is empty"
+            if (Test-EmptyDirectory -Path "$path") {
+                Write-Output "$tool does not exist at '$Path' or the directory is empty"
                 return
             }
 
             $consent = $force
             if (!$ENV:PESTER) {
-                $consent = $force -or $PSCmdlet.ShouldContinue($env:COMPUTERNAME, "Are you sure you want to uninstall Buildkit from $path?")
+                $consent = $force -or $PSCmdlet.ShouldContinue($env:COMPUTERNAME, "Are you sure you want to uninstall Buildkit from '$path'?")
             }
 
             if (!$consent) {
                 Throw "$tool uninstallation cancelled."
             }
 
-            Write-Warning "Uninstalling preinstalled $tool at the path $path"
+            Write-Warning "Uninstalling preinstalled $tool at the path '$path'.`n$WhatIfMessage"
             try {
-                Uninstall-BuildkitHelper -Path $path
+                Uninstall-BuildkitHelper -Path "$path" -Purge:$Purge
             }
             catch {
                 Throw "Could not uninstall $tool. $_"
@@ -389,35 +398,45 @@ function Uninstall-Buildkit {
 function Uninstall-BuildkitHelper {
     param(
         [parameter(Mandatory = $true, HelpMessage = "Buildkit path")]
-        [String]$Path
+        [String]$Path,
+
+        [parameter(HelpMessage = "Remove all program data for Buildkit")]
+        [Switch] $Purge
     )
 
-    if (Test-EmptyDirectory -Path $Path) {
-        Write-Error "Buildkit does not exist at $Path or the directory is empty."
+    if (Test-EmptyDirectory -Path "$Path") {
+        Write-Error "Buildkit does not exist at '$Path' or the directory is empty."
         return
     }
 
     try {
         if (Test-ServiceRegistered -Service 'Buildkitd') {
             Stop-BuildkitdService
-            Unregister-Buildkitd -BuildkitPath $Path
+            Unregister-Buildkitd -BuildkitPath "$Path"
         }
     }
     catch {
         Throw "Could not stop or unregister buildkitd service. $_"
     }
 
-    # Delete the buildkit key
-    Remove-Item -Path "HKLM:\SYSTEM\CurrentControlSet\Services\buildkit" -Recurse -Force -ErrorAction Ignore
-
     # Remove the folder where buildkit is installed and related folders
-    Remove-Item -Path $Path -Recurse -Force
+    Remove-Item -Path "$Path" -Recurse -Force
 
-    # Delete Buildkit programdata
-    Uninstall-ProgramFiles "$ENV:ProgramData\Buildkit"
+    if ($Purge) {
+        Write-Output "Purging Buildkit program data"
 
-    # Remove from env path
-    Remove-FeatureFromPath -Feature "buildkit"
+        # Delete the buildkit key
+        Write-Warning "Removing Buildkit registry key"
+        Remove-Item -Path "HKLM:\SYSTEM\CurrentControlSet\Services\buildkit" -Recurse -Force -ErrorAction Ignore
+
+        # Delete Buildkit programdata
+        Write-Warning "Removing Buildkit program data"
+        Uninstall-ProgramFiles "$ENV:ProgramData\Buildkit"
+
+        # Remove from env path
+        Write-Warning "Removing Buildkit from env path"
+        Remove-FeatureFromPath -Feature "buildkit"
+    }
 
     Write-Output "Successfully uninstalled buildkit."
 }
